pyproject.toml
[project]
name = "secunda-task"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "greenlet>=3.2.3",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
    "sqlparse>=0.5.3",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "ruff>=0.12.7",
]

[tool.ruff]
fix = true
unsafe-fixes = true
line-length = 120
lint.preview = true
lint.select = ["ALL"]
lint.ignore = [
    "ANN401", # Dynamically typed expressions (typing.Any) are disallowed
    "COM812", # Skip list trailing slash (because using isort black formatter)
    "CPY001", # Missing copyright notice at top of file
    "D100", # Missing docstring in public module
    "D104", # Missing docstring in public package
    "D203", # Incorrect docstring whitespace before parameters (because using isort black formatter)
    "D211", # No blank line after last docstring
    "D213", # Multi-line docstring summary should start at the second line
    "DOC402", # `yield` is not documented in docstring
    "I", # All imports (because using isort)
    "RUF002", # Cyrillic like latin symbols in docstring
    "RUF003", # Cyrillic like latin symbols in docstring
]
extend-exclude = ["alembic"]

[tool.mypy]
mypy_path = "src"
plugins = "pydantic.mypy"
incremental = true
warn_return_any = true
warn_unused_configs = true
show_error_context = true
disallow_subclassing_any = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true
follow_imports_for_stubs = true
explicit_package_bases = true

[tool.isort]
profile = "black"
lines_after_imports = 2
line_length = 120
force_sort_within_sections = true
float_to_top = true
